[backend] refactor some filter typings
