+++ opencti-platform/opencti-graphql/src/utils/stix-filtering/stix-filtering.ts
@@ -4,30 +4,12 @@
 import { INDICATOR_FILTER, } from '../filtering';
 import { getStixTesterFromFilterKey } from './stix-testers';
 import { testFilterGroup } from './boolean-logic-engine';
+import type { Filter, FilterGroup } from './filter-group';
 
 // TODO: changed by Cathia, to integrate properly with her
 const LABEL_FILTER = 'objectLabel';
 
-export type FilterMode = 'AND' | 'OR';
-export type FilterOperator = 'eq' | 'not_eq' | 'lt' | 'lte' | 'gt' | 'gte' | 'nil' | 'not_nil';
-
-export type Filter = {
-  // multiple keys possible (internal use, in streams it's not possible)
-  // TODO: it should probably be named keys, but that's another story.
-  key: string[] // name, entity_type, etc
-  mode: FilterMode
-  values: string[]
-  operator: FilterOperator
-};
-
-export type FilterGroup = {
-  mode: FilterMode
-  filters: Filter[]
-  filterGroups: FilterGroup[]
-};
 //----------------------------------------------------------------------------------------------------------
-
-export type TesterFunction = (stix: any, filter: Filter, useSideEventMatching?: boolean) => boolean;
 
 /**
  * Resolve some of the values (recursively) inside the filter group
